Refactor schema command
